+++ src/datatypes/health.ts
@@ -5,4 +5,3 @@ type Schemas = components["schemas"];
 export type Health = Schemas["Health"];
 export type HealthStatus = Schemas["HealthStatus"];
 export type ComponentHealth = Schemas["ComponentHealth"];
-export type CheckHealth = Schemas["CheckHealth"];